backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
transformers==4.44.2
torch>=2.2.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.8.2
pypdf==4.3.1
